requirements.txt
# API clients
openai==1.13.0
anthropic==0.8.0
requests==2.31.0

# GUI and automation
pyautogui==0.9.54
python-dotenv==1.0.0
pillow==10.1.0
tkinter>=8.6.12
pyperclip==1.8.2
selenium==4.15.2
pygetwindow==0.0.9
beautifulsoup4==4.12.2
pywin32==306; platform_system=="Windows"

# Base packages
typing-extensions==4.8.0
